+++ sickrage/providers/torrent/archetorrent.py
@@ -167,13 +167,12 @@ class ArcheTorrentProvider(TorrentProvider):
                     torrent_size = cells[size_index].get_text()
                     size = convert_size(torrent_size, -1)
 
-                    item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
-                            'leechers': leechers}
+                    results += [
+                        {'title': title, 'link': download_url, 'size': size, 'seeders': seeders, 'leechers': leechers}
+                    ]
 
                     if mode != 'RSS':
                         sickrage.app.log.debug('Found result: {}'.format(title))
-
-                    results.append(item)
                 except Exception:
                     sickrage.app.log.error('Failed parsing provider')