+++ sickrage/providers/torrent/limetorrents.py
@@ -132,13 +132,12 @@ class LimeTorrentsProvider(TorrentProvider):
 
                     size = convert_size(cells[labels.index('Size')].get_text(strip=True), -1)
 
-                    item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
-                            'leechers': leechers}
+                    results += [
+                        {'title': title, 'link': download_url, 'size': size, 'seeders': seeders, 'leechers': leechers}
+                    ]
 
                     if mode != 'RSS':
                         sickrage.app.log.debug("Found result: {}".format(title))
-
-                    results.append(item)
                 except Exception:
                     sickrage.app.log.error("Failed parsing provider")