+++ sickrage/providers/torrent/newpct.py
@@ -153,17 +153,12 @@ class NewpctProvider(TorrentProvider):
                 torrent_size = re.sub(r'Size: ([\d.]+).+([KMGT]B)', r'\1 \2', torrent_size)
                 size = convert_size(torrent_size, -1)
 
-                item = {
-                    'title': title,
-                    'link': download_url,
-                    'size': size,
-                    'seeders': seeders,
-                    'leechers': leechers,
-                }
-                if mode != 'RSS':
-                    sickrage.app.log.debug('Found result: {}'.format(title))
+                results += [
+                    {'title': title, 'link': download_url, 'size': size, 'seeders': seeders, 'leechers': leechers}
+                ]
 
-                    results.append(item)
+                if mode != 'RSS':
+                    sickrage.app.log.debug("Found result: {}".format(title))
             except Exception:
                 sickrage.app.log.error('Failed parsing provider')