+++ sickrage/providers/torrent/norbits.py
@@ -116,12 +116,12 @@ class NorbitsProvider(TorrentProvider):
                     leechers = try_int(item.pop('leechers', 0))
                     size = convert_size(item.pop('size', -1), -1)
 
-                    item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
-                            'leechers': leechers}
-                    if mode != 'RSS':
-                        sickrage.app.log.debug('Found result: {0}'.format(title))
+                    results += [
+                        {'title': title, 'link': download_url, 'size': size, 'seeders': seeders, 'leechers': leechers}
+                    ]
 
-                    results.append(item)
+                    if mode != 'RSS':
+                        sickrage.app.log.debug("Found result: {}".format(title))
                 except Exception:
                     sickrage.app.log.error("Failed parsing provider")