+++ sickrage/providers/torrent/nyaatorrents.py
@@ -110,12 +110,12 @@ class NyaaProvider(TorrentProvider):
 
                 size = convert_size(item['nyaa_size'], -1, units=['B', 'KIB', 'MIB', 'GIB', 'TIB', 'PIB'])
 
-                item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders, 'leechers': leechers}
+                results += [
+                    {'title': title, 'link': download_url, 'size': size, 'seeders': seeders, 'leechers': leechers}
+                ]
 
                 if mode != 'RSS':
-                    sickrage.app.log.debug('Found result: {}'.format(title))
-
-                results.append(item)
+                    sickrage.app.log.debug("Found result: {}".format(title))
             except Exception:
                 sickrage.app.log.error('Failed parsing provider')